requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
behave==1.2.6
behavex==2.0.1
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
configobj==5.0.8
csscompressor==0.9.5
distlib==0.3.7
filelock==3.12.2
htmlmin==0.1.12
identify==2.5.26
idna==3.4
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
parse==1.19.1
parse-type==0.6.2
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.3.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
urllib3==2.0.4
virtualenv==20.24.1
